dff-weather-skill: move changes from dff-template to dff-weather
