+++ response_selectors/convers_evaluation_based_selector/server.py
@@ -395,37 +395,37 @@ def select_response(
     best_human_attributes = best_candidate.get("human_attributes", {})
     best_bot_attributes = best_candidate.get("bot_attributes", {})
 
-    # greeting_spec = "this is an Alexa Prize Socialbot"
-    # if len(dialog["bot_utterances"]) == 0 and greeting_spec not in best_text:
-    #     # add greeting to the first bot uttr, if it's not already included
-    #     best_text = "Hi, " + greeting_spec + "! " + best_text
+    greeting_spec = "this is an Alexa Prize Socialbot"
+    if len(dialog["bot_utterances"]) == 0 and greeting_spec not in best_text:
+        # add greeting to the first bot uttr, if it's not already included
+        best_text = "Hi, " + greeting_spec + "! " + best_text
 
-    # while candidates[best_id]["text"] == "" or candidates[best_id]["confidence"] == 0.0:
-    #     curr_single_scores[int(best_id)] = 0.0
-    #     best_id = np.argmax(curr_single_scores)
-    #     best_skill_name = candidates[best_id]["skill_name"]
-    #     best_text = candidates[best_id]["text"]
-    #     best_confidence = candidates[best_id]["confidence"]
-    #     best_human_attributes = candidates[best_id].get("human_attributes", {})
-    #     best_bot_attributes = candidates[best_id].get("bot_attributes", {})
-    #     if sum(curr_single_scores) == 0.0:
-    #         break
+    while candidates[best_id]["text"] == "" or candidates[best_id]["confidence"] == 0.0:
+        curr_single_scores[int(best_id)] = 0.0
+        best_id = np.argmax(curr_single_scores)
+        best_skill_name = candidates[best_id]["skill_name"]
+        best_text = candidates[best_id]["text"]
+        best_confidence = candidates[best_id]["confidence"]
+        best_human_attributes = candidates[best_id].get("human_attributes", {})
+        best_bot_attributes = candidates[best_id].get("bot_attributes", {})
+        if sum(curr_single_scores) == 0.0:
+            break
 
-    # if dialog["human"]["profile"].get("name", False) and best_skill_name != "personal_info_skill":
-    #     name = dialog["human"]["profile"].get("name", False)
-    #     if len(dialog["bot_utterances"]) >= 1:
-    #         if re.search(r"\b" + name + r"\b", dialog["bot_utterances"][-1]["text"]):
-    #             pass
-    #         else:
-    #             if random.random() <= CALL_BY_NAME_PROBABILITY:
-    #                 best_text = f"{name}, {best_text}"
-    #     else:
-    #         # if dialog is just started (now it's impossible)
-    #         if random.random() <= CALL_BY_NAME_PROBABILITY:
-    #             best_text = f"{name}, {best_text}"
+    if dialog["human"]["profile"].get("name", False) and best_skill_name != "personal_info_skill":
+        name = dialog["human"]["profile"].get("name", False)
+        if len(dialog["bot_utterances"]) >= 1:
+            if re.search(r"\b" + name + r"\b", dialog["bot_utterances"][-1]["text"]):
+                pass
+            else:
+                if random.random() <= CALL_BY_NAME_PROBABILITY:
+                    best_text = f"{name}, {best_text}"
+        else:
+            # if dialog is just started (now it's impossible)
+            if random.random() <= CALL_BY_NAME_PROBABILITY:
+                best_text = f"{name}, {best_text}"
 
-    # if dialog["human_utterances"][-1]["text"] == "/get_dialog_id":
-    #     best_text = "Your dialog's id: " + str(dialog["dialog_id"])
+    if dialog["human_utterances"][-1]["text"] == "/get_dialog_id":
+        best_text = "Your dialog's id: " + str(dialog["dialog_id"])
 
     return best_skill_name, best_text, best_confidence, best_human_attributes, best_bot_attributes
