+++ skills/dff_template_skill/scenario/main.py
@@ -1,73 +1,115 @@
-import dff.conditions as cnd
+import logging
+import re
+
+from dff.core.keywords import PROCESSING, TRANSITIONS, GRAPH, RESPONSE, GLOBAL_TRANSITIONS
 from dff.core import Actor
-from dff.core.keywords import GLOBAL_TRANSITIONS, GRAPH, PROCESSING, RESPONSE, TRANSITIONS
+import dff.conditions as cnd
+import dff.transitions as trn
 
 import common.dff.integration.condition as int_cnd
 import common.dff.integration.processing as int_prs
-from common.constants import CAN_CONTINUE_SCENARIO
-from common.weather import ASK_WEATHER_SKILL_PHRASE
-from scenario.condition import (
-    chat_about_weather_condition,
-    forecast_intent_condition,
-    forecast_requested_condition,
-    homeland_forecast_requested_condition,
-    request_with_location_condition,
-)
-from scenario.constants import HIGH_CONF, ZERO_CONF
-from scenario.processing import location_request_processing
-from scenario.response import activity_answer_response, activity_question_response, forecast_response
+import common.dff.integration.response as int_rsp
 
+
+import common.constants as common_constants
+
+from . import condition as loc_cnd
+from . import response as loc_rsp
+
+# from . import processing as loc_prs
+
+logger = logging.getLogger(__name__)
+
+# First of all, to create a dialog agent, we need to create a dialog script.
+# Below, `flows` is the dialog script.
+# A dialog script is a flow dictionary that can contain multiple flows .
+# Flows are needed in order to divide a dialog into sub-dialogs and process them separately.
+# For example, the separation can be tied to the topic of the dialog.
+# In our example, there is one flow called greeting_flow.
+
+# Inside each flow, we can describe a sub-dialog using keyword `GRAPH` from dff.core.keywords module.
+# Here we can also use keyword `GLOBAL_TRANSITIONS`, which we have considered in other examples.
+
+# `GRAPH` describes a sub-dialog using linked nodes, each node has the keywords `RESPONSE` and `TRANSITIONS`.
+
+# `RESPONSE` - contains the response that the dialog agent will return when transitioning to this node.
+# `TRANSITIONS` - describes transitions from the current node to other nodes.
+# `TRANSITIONS` are described in pairs:
+#      - the node to which the agent will perform the transition
+#      - the condition under which to make the transition
+
+std_prs = [int_prs.set_confidence(1.0), int_prs.set_can_continue()]
 flows = {
     "global": {
+        GLOBAL_TRANSITIONS: {
+            ("greeting", "node1"): cnd.regexp(r"\bhi\b"),
+        },
         GRAPH: {
             "start": {
                 RESPONSE: "",
-                TRANSITIONS: {
-                    ("weather", "forecast"): cnd.any(
-                        [
-                            homeland_forecast_requested_condition,
-                            cnd.all(
-                                [
-                                    request_with_location_condition,
-                                    cnd.any([forecast_requested_condition, forecast_intent_condition]),
-                                ]
-                            ),
-                        ]
-                    ),
-                    ("weather", "location_request"): cnd.any([forecast_requested_condition, forecast_intent_condition]),
-                    ("weather", "continue_question"): chat_about_weather_condition,
-                },
+                TRANSITIONS: {("greeting", "node1"): cnd.true},
             },
             "fallback": {
-                RESPONSE: "Oops",
-                PROCESSING: [int_prs.set_confidence(ZERO_CONF)],
+                RESPONSE: "Ooops",
+                TRANSITIONS: {
+                    trn.previous(): cnd.regexp(r"previous", re.IGNORECASE),
+                    trn.repeat(0.2): cnd.true,
+                },
             },
         },
     },
-    "weather": {
+    "greeting": {
         GRAPH: {
-            "forecast": {RESPONSE: forecast_response, TRANSITIONS: {"activity_question": cnd.true}},
-            "location_request": {
-                RESPONSE: "Hmm. Which particular city would you like a weather forecast for?",
-                PROCESSING: [location_request_processing],
-                TRANSITIONS: {"forecast": cnd.true},
+            "node1": {
+                RESPONSE: int_rsp.multi_response(replies=["Hi, how are you?", "Hi, what's up?"]),  # several hypothesis
+                PROCESSING: std_prs + [int_prs.save_slots_to_ctx({"topic": "science", "user_name": "Gordon Freeman"})],
+                TRANSITIONS: {"node2": cnd.regexp(r"how are you", re.IGNORECASE)},
             },
-            "continue_question": {
-                RESPONSE: ASK_WEATHER_SKILL_PHRASE,
-                PROCESSING: [int_prs.set_confidence(HIGH_CONF)],
+            "node2": {
+                RESPONSE: loc_rsp.example_response("Good. What do you want to talk about?"),
+                # loc_rsp.example_response is just for an example, you can use just str without example_response func
+                PROCESSING: std_prs,
+                TRANSITIONS: {"node3": loc_cnd.example_lets_talk_about()},
+            },
+            "node3": {
+                RESPONSE: "Sorry, I can not talk about that now. Maybe late. Do you like {topic}?",
+                PROCESSING: std_prs + [int_prs.fill_responses_by_slots()],
                 TRANSITIONS: {
-                    "forecast": cnd.all([int_cnd.is_yes_vars, request_with_location_condition]),
-                    "location_request": int_cnd.is_yes_vars,
+                    "node4": int_cnd.is_yes_vars,
+                    "node5": int_cnd.is_no_vars,
+                    "node6": int_cnd.is_do_not_know_vars,
+                    "node7": cnd.true,  # it will be chosen if other conditions are False
                 },
             },
-            "activity_question": {
-                RESPONSE: activity_question_response,
-                PROCESSING: [int_prs.set_can_continue(CAN_CONTINUE_SCENARIO)],
-                TRANSITIONS: {"activity_answer": cnd.true},
+            "node4": {
+                RESPONSE: "I like {topic} too, {user_name}",
+                PROCESSING: std_prs + [int_prs.fill_responses_by_slots()],
+                TRANSITIONS: {("node7", 0.1): cnd.true},
             },
-            "activity_answer": {RESPONSE: activity_answer_response},
+            "node5": {
+                RESPONSE: "I do not like {topic} too, {user_name}",
+                PROCESSING: std_prs + [int_prs.fill_responses_by_slots()],
+                TRANSITIONS: {("node7", 0.1): cnd.true},
+            },
+            "node6": {
+                RESPONSE: "I have no opinion about {topic} too, {user_name}",
+                PROCESSING: std_prs + [int_prs.fill_responses_by_slots()],
+                TRANSITIONS: {("node7", 0.1): cnd.true},
+            },
+            "node7": {
+                RESPONSE: int_rsp.multi_response(
+                    replies=["bye", "goodbye"],
+                    confidences=[1.0, 0.5],
+                    hype_attr=[
+                        {"can_continue": common_constants.MUST_CONTINUE},  # for the first hyp
+                        {"can_continue": common_constants.CAN_CONTINUE_SCENARIO},  # for the second hyp
+                    ],
+                ),
+                PROCESSING: [int_prs.set_can_continue()],
+            },
         }
     },
 }
+
 
 actor = Actor(flows, start_node_label=("global", "start"), fallback_node_label=("global", "fallback"))
