+++ skills/dff_template_skill/tests/lets_talk_out.json
@@ -1,15 +1,15 @@
 [
     [
-        "It is scattered clouds, temperature is around 67.6 Fahrenheits in London. Wind speed is about 1.0 meters per second. What weather do you prefer? Warm, cold, rain, snow, hot?",
-        1.0,
+        "Hi, how are you?",
+        0.85,
         {
             "dff_template_skill_state": {
                 "shared_memory": {},
                 "previous_human_utter_index": 0,
                 "history": {
                     "0": [
-                        "weather",
-                        "forecast"
+                        "greeting",
+                        "node1"
                     ]
                 },
                 "current_turn_dff_suspended": false,
@@ -17,15 +17,15 @@
                     "id": "fc82c8b9-1a70-4a64-bfd5-d37142d6d52a",
                     "node_labels": {
                         "0": [
-                            "weather",
-                            "forecast"
+                            "greeting",
+                            "node1"
                         ]
                     },
                     "requests": {
-                        "0": "Tell me weather in London."
+                        "0": "hi."
                     },
                     "responses": {
-                        "0": "It is scattered clouds, temperature is around 67.6 Fahrenheits in London. Wind speed is about 1.0 meters per second. What weather do you prefer? Warm, cold, rain, snow, hot?"
+                        "0": "Hi, how are you?"
                     },
                     "previous_index": 0,
                     "current_index": 0,
@@ -43,7 +43,7 @@
         },
         {},
         {
-            "can_continue": "must"
+            "can_continue": "can"
         }
     ]
 ]
