+++ skills/dff_weather_skill/requirements.txt
@@ -1,2 +1,3 @@
 click==7.1.2
 nltk==3.5
+ahocorapy==1.5.2
