+++ skills/dff_weather_skill/scenario/constants.py
