+++ skills/dff_weather_skill/test.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python test_server.py
