+++ skills/dff_weather_skill/test_server.py
@@ -22,7 +22,7 @@ def run_test(handler):
         is_equal_flag, msg = test_utils.compare_structs(out_data[test_name], hypothesis, ignored_keys=["id"])
         if msg and len(msg.split("`")) == 5:
             _, ground_truth_text, _, hypothesis_text, _ = msg.split("`")
-            is_equal_flag, ratio = test_utils.compare_text(ground_truth_text, hypothesis_text, 0.80)
+            is_equal_flag, ratio = test_utils.compare_text(ground_truth_text, hypothesis_text, 0.75)
             if not is_equal_flag:
                 msg = f"{msg} ratio = {ratio}"
         assert is_equal_flag, msg
