+++ skills/dff_weather_skill/tests/activity_question_in.json
@@ -1032,7 +1032,7 @@
                             ]
                         }
                     },
-                    "active_skill": "dff_template_skill"
+                    "active_skill": "dff_weather_skill"
                 },
                 {
                     "text": "I prefer cold weather too! Let me guess you like ice-skating?",
@@ -1057,12 +1057,12 @@
                             ]
                         }
                     },
-                    "active_skill": "dff_template_skill"
+                    "active_skill": "dff_weather_skill"
                 }
             ]
         }
     ],
-    "dff_template_skill_state_batch": [
+    "dff_weather_skill_state_batch": [
         {
             "context": {
                 "current_index": 1,
