+++ skills/dff_weather_skill/tests/activity_question_out.json
@@ -1,13 +1,13 @@
 [
     [
-        "I prefer cold weather too! Let me guess you like ice-skating?",
+        "Society is like a large piece of frozen water; and skating well is the great art of social life. Do you have pets?",
         0.999,
         {
-            "dff_template_skill_state": {
+            "dff_weather_skill_state": {
                 "shared_memory": {
-                    "preferred_weather": "cold"
+                    "preferred_weather": ""
                 },
-                "previous_human_utter_index": 1,
+                "previous_human_utter_index": 2,
                 "history": {
                     "0": [
                         "weather",
@@ -16,31 +16,35 @@
                     "1": [
                         "weather",
                         "activity_question"
+                    ],
+                    "2": [
+                        "weather",
+                        "activity_answer"
                     ]
                 },
                 "current_turn_dff_suspended": false,
                 "context": {
                     "id": "fc82c8b9-1a70-4a64-bfd5-d37142d6d52a",
                     "node_labels": {
-                        "0": [
-                            "weather",
-                            "forecast"
-                        ],
                         "1": [
                             "weather",
                             "activity_question"
+                        ],
+                        "2": [
+                            "weather",
+                            "activity_answer"
                         ]
                     },
                     "requests": {
-                        "0": "Tell me weather in London.",
-                        "1": "I prefer cold weather."
+                        "1": "I prefer cold weather.",
+                        "2": "Yes"
                     },
                     "responses": {
-                        "0": "It is broken clouds, temperature is around 71.0 degrees Fahrenheit in London. Wind speed is about 0.5 meters per second. What weather do you prefer? Warm, cold, rain, snow, hot?",
-                        "1": "I prefer cold weather too! Let me guess you like ice-skating?"
+                        "1": "I prefer cold weather too! Let me guess you like ice-skating?",
+                        "2": "Society is like a large piece of frozen water; and skating well is the great art of social life. Do you have pets?"
                     },
-                    "previous_index": 1,
-                    "current_index": 1,
+                    "previous_index": 2,
+                    "current_index": 2,
                     "misc": {},
                     "validation": false
                 }
@@ -49,7 +53,11 @@
                 "cross_states": {},
                 "cross_links": {}
             },
-            "used_links": {},
+            "used_links": {
+                "dff_animals_skill": [
+                    "Do you have pets?"
+                ]
+            },
             "age_group": "unknown",
             "disliked_skills": []
         },
