+++ skills/dff_weather_skill/tests/preferred_weather_out.json
@@ -1,13 +1,13 @@
 [
     [
-        "Society is like a large piece of frozen water; and skating well is the great art of social life. Do you have pets?",
+        "I prefer cold weather too! Let me guess you like ice-skating?",
         0.999,
         {
-            "dff_template_skill_state": {
+            "dff_weather_skill_state": {
                 "shared_memory": {
-                    "preferred_weather": ""
+                    "preferred_weather": "cold"
                 },
-                "previous_human_utter_index": 2,
+                "previous_human_utter_index": 1,
                 "history": {
                     "0": [
                         "weather",
@@ -16,35 +16,31 @@
                     "1": [
                         "weather",
                         "activity_question"
-                    ],
-                    "2": [
-                        "weather",
-                        "activity_answer"
                     ]
                 },
                 "current_turn_dff_suspended": false,
                 "context": {
                     "id": "fc82c8b9-1a70-4a64-bfd5-d37142d6d52a",
                     "node_labels": {
+                        "0": [
+                            "weather",
+                            "forecast"
+                        ],
                         "1": [
                             "weather",
                             "activity_question"
-                        ],
-                        "2": [
-                            "weather",
-                            "activity_answer"
                         ]
                     },
                     "requests": {
-                        "1": "I prefer cold weather.",
-                        "2": "Yes"
+                        "0": "Tell me weather in London.",
+                        "1": "I prefer cold weather."
                     },
                     "responses": {
-                        "1": "I prefer cold weather too! Let me guess you like ice-skating?",
-                        "2": "Society is like a large piece of frozen water; and skating well is the great art of social life. Do you have pets?"
+                        "0": "It is broken clouds, temperature is around 71.0 degrees Fahrenheit in London. Wind speed is about 0.5 meters per second. What weather do you prefer? Warm, cold, rain, snow, hot?",
+                        "1": "I prefer cold weather too! Let me guess you like ice-skating?"
                     },
-                    "previous_index": 2,
-                    "current_index": 2,
+                    "previous_index": 1,
+                    "current_index": 1,
                     "misc": {},
                     "validation": false
                 }
@@ -53,11 +49,7 @@
                 "cross_states": {},
                 "cross_links": {}
             },
-            "used_links": {
-                "dff_animals_skill": [
-                    "Do you have pets?"
-                ]
-            },
+            "used_links": {},
             "age_group": "unknown",
             "disliked_skills": []
         },
