+++ skills/dff_weather_skill/tests/request_with_city_out.json
@@ -1,9 +1,9 @@
 [
     [
-        "Currently I have problems with getting weather in mockcity. But I would guess... Potato storm may wonder you. . And the temperature may be around 3 degrees Fahrenheit. Would you enjoy it?. What weather do you prefer? Warm, cold, rain, snow, hot?",
+        "It is scattered clouds, temperature is around 67.6 Fahrenheits in London. Wind speed is about 1.0 meters per second. What weather do you prefer? Warm, cold, rain, snow, hot?",
         1.0,
         {
-            "dff_template_skill_state": {
+            "dff_weather_skill_state": {
                 "shared_memory": {},
                 "previous_human_utter_index": 0,
                 "history": {
@@ -22,10 +22,10 @@
                         ]
                     },
                     "requests": {
-                        "0": "Tell me weather in mockcity."
+                        "0": "Tell me weather in London."
                     },
                     "responses": {
-                        "0": "Currently r in mockcity. But I would guess... Potato storm may wonder you. . And the temperature may be around 3 degrees Fahrenheit. Would you enjoy it?. What weather do you prefer? Warm, cold, rain, snow, hot?"
+                        "0": "It is scattered clouds, temperature is around 67.6 Fahrenheits in London. Wind speed is about 1.0 meters per second. What weather do you prefer? Warm, cold, rain, snow, hot?"
                     },
                     "previous_index": 0,
                     "current_index": 0,
