+++ skills/dff_weather_skill/tests/request_with_nonexistent_city_out.json
@@ -1,28 +1,31 @@
 [
     [
-        "Hmm. Which particular city would you like a weather forecast for?",
+        "Currently I have problems with getting weather in mockcity. But I would guess... Potato storm may wonder you. . And the temperature may be around 3 degrees Fahrenheit. Would you enjoy it?. What weather do you prefer? Warm, cold, rain, snow, hot?",
         1.0,
         {
-            "dff_template_skill_state": {
+            "dff_weather_skill_state": {
                 "shared_memory": {},
                 "previous_human_utter_index": 0,
                 "history": {
                     "0": [
                         "weather",
-                        "location_request"
+                        "forecast"
                     ]
                 },
                 "current_turn_dff_suspended": false,
                 "context": {
                     "id": "fc82c8b9-1a70-4a64-bfd5-d37142d6d52a",
                     "node_labels": {
-                        "0": ["weather", "location_request"]
+                        "0": [
+                            "weather",
+                            "forecast"
+                        ]
                     },
                     "requests": {
-                        "0": "Tell me weather."
+                        "0": "Tell me weather in mockcity."
                     },
                     "responses": {
-                        "0": "Hmm. Which particular city would you like a weather forecast for?"
+                        "0": "Currently r in mockcity. But I would guess... Potato storm may wonder you. . And the temperature may be around 3 degrees Fahrenheit. Would you enjoy it?. What weather do you prefer? Warm, cold, rain, snow, hot?"
                     },
                     "previous_index": 0,
                     "current_index": 0,
