+++ skills/dff_weather_skill/tests/request_without_city_out.json
@@ -1,0 +1,46 @@
+[
+    [
+        "Hmm. Which particular city would you like a weather forecast for?",
+        1.0,
+        {
+            "dff_weather_skill_state": {
+                "shared_memory": {},
+                "previous_human_utter_index": 0,
+                "history": {
+                    "0": [
+                        "weather",
+                        "location_request"
+                    ]
+                },
+                "current_turn_dff_suspended": false,
+                "context": {
+                    "id": "fc82c8b9-1a70-4a64-bfd5-d37142d6d52a",
+                    "node_labels": {
+                        "0": ["weather", "location_request"]
+                    },
+                    "requests": {
+                        "0": "Tell me weather."
+                    },
+                    "responses": {
+                        "0": "Hmm. Which particular city would you like a weather forecast for?"
+                    },
+                    "previous_index": 0,
+                    "current_index": 0,
+                    "misc": {},
+                    "validation": false
+                }
+            },
+            "dff_shared_state": {
+                "cross_states": {},
+                "cross_links": {}
+            },
+            "used_links": {},
+            "age_group": "unknown",
+            "disliked_skills": []
+        },
+        {},
+        {
+            "can_continue": "must"
+        }
+    ]
+]
